+++ packages/cloud-cognitive/src/components/AddSelect/AddSelectSidebar.js
@@ -6,12 +6,13 @@
 //
 
 import React from 'react';
-import { Tag, Accordion, AccordionItem, Button } from 'carbon-components-react';
-import { SubtractAlt32 } from '@carbon/icons-react';
+import { Tag, Accordion, AccordionItem } from 'carbon-components-react';
 import PropTypes from 'prop-types';
 import { NoDataEmptyState } from '../../components/EmptyStates/NoDataEmptyState';
 import { pkg } from '../../settings';
 import { AddSelectMetaPanel } from './AddSelectMetaPanel';
+
+const blockClass = `${pkg.prefix}--add-select__sidebar`;
 const componentName = 'AddSelectSidebar';
 
 export let AddSelectSidebar = ({
@@ -25,27 +26,25 @@ export let AddSelectSidebar = ({
   multiSelection,
   noSelectionDescription,
   noSelectionTitle,
-  removeIconDescription,
   setDisplayMetaPanel,
-  setMultiSelection,
 }) => {
-  const blockClass = `${pkg.prefix}--add-select__sidebar`;
   const hasModifiers = modifiers?.options?.length > 0;
   const hasSelections = multiSelection.length > 0;
 
-  const handleItemRemove = (id) => {
-    const newSelections = multiSelection.filter((v) => v !== id);
-    setMultiSelection(newSelections);
-  };
-
-  const sidebarItems = multiSelection.reduce((acc, cur) => {
-    const selectedItem = items.find((item) => item.id === cur);
+  const getNewItem = (item) => {
     // certain properties should not be displayed in the sidebar
     // eslint-disable-next-line no-unused-vars
-    const { meta, icon, avatar, ...newItem } = selectedItem;
-    acc.push(newItem);
-    return acc;
-  }, []);
+    const { meta, icon, avatar, ...newItem } = item;
+    return newItem;
+  };
+
+  const sidebarItems = multiSelection.map((selectionId) => {
+    if (Array.isArray(items)) {
+      const selectedItem = items.find((item) => item.id === selectionId);
+      return getNewItem(selectedItem);
+    }
+    return getNewItem(items[selectionId]);
+  });
 
   const getTitle = (item) => (
     <div className={`${blockClass}-accordion-title`}>
@@ -64,15 +63,6 @@ export let AddSelectSidebar = ({
           }
         </div>
       )}
-      <Button
-        renderIcon={SubtractAlt32}
-        iconDescription={removeIconDescription}
-        hasIconOnly
-        onClick={() => handleItemRemove(item.id)}
-        kind="ghost"
-        className={`${blockClass}-item-remove-button`}
-        size="sm"
-      />
     </div>
   );
 
@@ -126,15 +116,13 @@ AddSelectSidebar.propTypes = {
   closeIconDescription: PropTypes.string,
   displayMetalPanel: PropTypes.object,
   influencerTitle: PropTypes.string,
-  items: PropTypes.array,
+  items: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
   metaPanelTitle: PropTypes.string,
   modifiers: PropTypes.object,
   multiSelection: PropTypes.array,
   noSelectionDescription: PropTypes.string,
   noSelectionTitle: PropTypes.string,
-  removeIconDescription: PropTypes.string,
   setDisplayMetaPanel: PropTypes.func,
-  setMultiSelection: PropTypes.func,
 };
 
 AddSelectSidebar.displayName = componentName;