+++ packages/cloud-cognitive/src/components/CreateTearsheet/preview-components/MultiStepWithIntro.js
@@ -194,11 +194,7 @@ export const MultiStepWithIntro = ({
                 }}
                 invalid={isInvalid}
                 invalidText="This is a required field"
-                onBlur={() => {
-                  if (!stepOneTextInputValue.length) {
-                    setIsInvalid(true);
-                  }
-                }}
+                onBlur={() => {}}
               />
               <TextInput
                 labelText="Topic description (optional)"