+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid.stories.js
@@ -19,6 +19,7 @@ import {
   Restart16,
   Download16,
   Filter16,
+  Add16,
 } from '@carbon/icons-react';
 import { DataTable, Button, Pagination } from 'carbon-components-react';
 import {
@@ -35,18 +36,24 @@ import {
   useDisableSelectRows,
   useCustomizeColumns,
   useSelectAllWithToggle,
+  useColumnCenterAlign,
   useStickyColumn,
   useActionsColumn,
 } from '.';
+
 import {
-  /*StickyActionsColumn,*/ CustomizeColumnStory,
+  CustomizeColumnStory,
   RowSizeDropdownStory,
   SelectAllWitHToggle,
   LeftPanelStory,
 } from './Datagrid.stories';
 import mdx from './Datagrid.mdx';
 
+import { pkg } from '../../settings';
+import cx from 'classnames';
+
 import styles from './_storybook-styles.scss';
+import { SidePanel } from '../SidePanel';
 
 export default {
   title: getStoryTitle(Datagrid.displayName),
@@ -62,6 +69,8 @@ export default {
     },
   },
 };
+
+const blockClass = `${pkg.prefix}--datagrid`;
 
 const Wrapper = ({ children }) => (
   <div
@@ -209,6 +218,49 @@ export const InitialLoad = () => {
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
+export const WithHeader = () => {
+  const columns = React.useMemo(() => defaultHeader, []);
+  const [data] = useState(makeData(11));
+  const gridTitle = 'Data table title';
+  const gridDescription = 'Additional information if needed';
+  const datagridState = useDatagrid({
+    columns,
+    data,
+    gridTitle,
+    gridDescription,
+    initialState: {
+      pageSize: 10,
+      pageSizes: [5, 10, 25, 50],
+    },
+    DatagridActions,
+    DatagridPagination,
+  });
+
+  return <Datagrid datagridState={{ ...datagridState }} />;
+};
+
+export const WithDenseHeader = () => {
+  const columns = React.useMemo(() => defaultHeader, []);
+  const [data] = useState(makeData(11));
+  const gridTitle = 'Data table title';
+  const gridDescription = 'Additional information if needed';
+  const datagridState = useDatagrid({
+    columns,
+    data,
+    gridTitle,
+    gridDescription,
+    initialState: {
+      pageSize: 10,
+      pageSizes: [5, 10, 25, 50],
+    },
+    useDenseHeader: true,
+    DatagridActions,
+    DatagridPagination,
+  });
+
+  return <Datagrid datagridState={{ ...datagridState }} />;
+};
+
 export const InfiniteScroll = () => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data, setData] = useState(makeData(0));
@@ -239,7 +291,7 @@ export const InfiniteScroll = () => {
 
   return (
     <Wrapper>
-      <Datagrid datagridState={{ ...datagridState }} />;
+      <Datagrid datagridState={{ ...datagridState }} />
     </Wrapper>
   );
 };
@@ -306,7 +358,6 @@ export const NestedRows = () => {
 };
 export const ExpandedRow = () => {
   const expansionRenderer = ({ row }) => <div>Content for {row.id}</div>;
-
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10));
   const datagridState = useDatagrid(
@@ -360,16 +411,88 @@ export const NestedTable = () => {
 export const ClickableRow = () => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10));
+  const [openSidePanel, setOpenSidePanel] = useState(false);
+  const [rowData, setRowData] = useState({});
   const datagridState = useDatagrid(
     {
       columns,
       data,
-      onRowClick: (row) => alert(`Clicked ${row.id}`),
+      onRowClick: (row) => {
+        setOpenSidePanel(true);
+        setRowData(row);
+      },
     },
     useOnRowClick
   );
+  return (
+    <div
+      className={cx(
+        openSidePanel
+          ? `page-content-wrapper side-panel-open`
+          : 'page-content-wrapper'
+      )}
+    >
+      <Datagrid datagridState={{ ...datagridState }} />
+      <SidePanel
+        selectorPageContent={true && '.page-content-wrapper'} // Only if SlideIn
+        open={openSidePanel}
+        onRequestClose={() => setOpenSidePanel(false)}
+        size={'sm'}
+        title={'Title'}
+        slideIn
+      >
+        <DataTableSidePanelContent rowData={rowData && rowData.original} />
+      </SidePanel>
+    </div>
+  );
+};
 
-  return <Datagrid datagridState={{ ...datagridState }} />;
+const DataTableSidePanelContent = (selectedRowValues) => {
+  const { rowData } = selectedRowValues;
+
+  const SidePanelSectionContent = ({ rowData, columns, sectionTitle }) => {
+    const finalData = columns.map((item) => Object.entries(rowData)[item]);
+    return (
+      <div className={`${blockClass}__side-panel-sections`}>
+        <h5 className={`${blockClass}__side-panel-section-header`}>
+          {sectionTitle}
+        </h5>
+        {finalData.map(([label, value], index) => {
+          return (
+            <div
+              key={index}
+              className={`${blockClass}__side-panel-section-inner`}
+            >
+              <div className={`${blockClass}__side-panel-label-text`}>
+                {label} :
+              </div>
+              <div className={`${blockClass}__side-panel-value`}>{value}</div>
+            </div>
+          );
+        })}
+      </div>
+    );
+  };
+
+  return (
+    <div className={`${blockClass}__side-panel-content`}>
+      <SidePanelSectionContent
+        sectionTitle="Section title"
+        rowData={rowData && rowData}
+        columns={[0]}
+      />
+      <SidePanelSectionContent
+        sectionTitle="Personal details"
+        rowData={rowData && rowData}
+        columns={[1, 2, 3, 4]}
+      />
+      <SidePanelSectionContent
+        sectionTitle="Section title"
+        rowData={rowData && rowData}
+        columns={[5, 6, 7, 8, 9, 10, 11, 12]}
+      />
+    </div>
+  );
 };
 
 export const IsHoverOnRow = () => {
@@ -426,7 +549,7 @@ export const RadioSelect = () => {
       data,
       hideSelectAll: true,
       radio: true,
-      onRadioSelect: (row) => console.log(row),
+      onRadioSelect: (row) => console.log(`Row clicked: ${row.id}`),
       initialState: {
         selectedRowIds: {
           3: true,
@@ -497,6 +620,37 @@ export const RightAlignedColumns = () => {
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
 
+export const CenterAlignedColumns = () => {
+  const columns = React.useMemo(
+    () => [
+      ...defaultHeader.slice(0, 3),
+      {
+        Header: 'Age',
+        accessor: 'age',
+        centerAlignedColumn: true,
+      },
+
+      {
+        Header: 'Visit',
+        accessor: 'visits',
+        centerAlignedColumn: true,
+      },
+    ],
+    []
+  );
+
+  const [data] = useState(makeData(10));
+  const datagridState = useDatagrid(
+    {
+      columns,
+      data,
+    },
+    useColumnCenterAlign
+  );
+
+  return <Datagrid datagridState={{ ...datagridState }} />;
+};
+
 const DatagridActions = (datagridState) => {
   const {
     selectedFlatRows,
@@ -504,6 +658,7 @@ const DatagridActions = (datagridState) => {
     CustomizeColumnsButton,
     RowSizeDropdown,
     rowSizeDropdownProps,
+    useDenseHeader,
   } = datagridState;
   const downloadCsv = () => {
     alert('Downloading...');
@@ -524,8 +679,44 @@ const DatagridActions = (datagridState) => {
     },
   };
   return (
-    isNothingSelected && (
-      <React.Fragment>
+    isNothingSelected &&
+    (useDenseHeader && useDenseHeader ? (
+      <TableToolbarContent size="sm">
+        <div style={style}>
+          <Button
+            kind="ghost"
+            hasIconOnly
+            tooltipPosition="bottom"
+            renderIcon={Download16}
+            iconDescription={'Download CSV'}
+            onClick={downloadCsv}
+          />
+        </div>
+        <div style={style}>
+          <Button
+            kind="ghost"
+            hasIconOnly
+            tooltipPosition="bottom"
+            renderIcon={Filter16}
+            iconDescription={'Left panel'}
+            onClick={leftPanelClick}
+          />
+        </div>
+        <RowSizeDropdown {...rowSizeDropdownProps} />
+        <div style={style} className={`${blockClass}__toolbar-divider`}>
+          <Button kind="ghost" renderIcon={Add16} iconDescription={'Action'}>
+            Ghost button
+          </Button>
+        </div>
+
+        {CustomizeColumnsButton && (
+          <div style={style}>
+            <CustomizeColumnsButton />
+          </div>
+        )}
+      </TableToolbarContent>
+    ) : (
+      <>
         <Button
           kind="ghost"
           hasIconOnly
@@ -569,8 +760,8 @@ const DatagridActions = (datagridState) => {
             </div>
           )}
         </TableToolbarContent>
-      </React.Fragment>
-    )
+      </>
+    ))
   );
 };
 
@@ -632,6 +823,7 @@ export const CustomizingColumns = () => {
   const datagridState = useDatagrid(
     {
       columns,
+      className: `c4p--datagrid__hidden--columns`,
       data,
       initialState: {
         hiddenColumns: ['age'],
@@ -721,14 +913,16 @@ export const LeftPanel = () => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10));
   const datagridState = useDatagrid({
-    columns,
-    data,
     leftPanel: {
       isOpen: true, // this toggling will happen from datagridActions.
       panelContent: (
-        <div>Panel content will go here along with any button interactions</div>
+        <div className={`${blockClass}__panel-content`}>
+          Panel content will go here along with any button interactions
+        </div>
       ),
     },
+    columns,
+    data,
     DatagridActions,
     DatagridBatchActions,
   });