+++ packages/cloud-cognitive/src/components/TagSet/TagSet.stories.js
@@ -192,6 +192,15 @@ export const ManyTags = prepareStory(Template, {
   },
 });
 
+export const MultilineTags = prepareStory(Template, {
+  args: {
+    tags: manyTags,
+    containerWidth: 500,
+    multiline: true,
+    ...overflowAndModalStrings,
+  },
+});
+
 export const HundredsOfTags = prepareStory(Template, {
   args: {
     tags: hundredsOfTags,